requirements.txt
asgiref==3.4.1
astroid==2.8.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.45.0
django-phonenumber-field==5.2.0
gunicorn==20.1.0
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.1
phonenumbers==8.12.31
Pillow==8.3.2
platformdirs==2.3.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
pylint==2.11.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
